REF: Move signed URL generator mock so it's importable externally
