+++ tests/cloud/storage/test_client.py
@@ -9,11 +9,11 @@ import google.cloud.exceptions
 import requests
 
 from octue.cloud import storage
+from octue.cloud.emulators import mock_generate_signed_url
 from octue.cloud.storage import GoogleCloudStorageClient
 from octue.exceptions import CloudStorageBucketNotFound
 from tests import TEST_BUCKET_NAME
 from tests.base import BaseTestCase
-from tests.mocks import mock_generate_signed_url
 
 
 class TestGoogleCloudStorageClient(BaseTestCase):
